+++ src/Services/ServiceBuilder.php
@@ -44,6 +44,7 @@ use Bitrix24\SDK\Legacy\LegacyServiceBuilder;
 use Bitrix24\SDK\Services\Lists\ListsServiceBuilder;
 use Bitrix24\SDK\Services\MailService\MailServiceServiceBuilder;
 use Bitrix24\SDK\Services\Messageservice\MessageserviceServiceBuilder;
+use Bitrix24\SDK\Services\Note\NoteServiceBuilder;
 use Bitrix24\SDK\Services\Rest\RestServiceBuilder;
 use Bitrix24\SDK\Services\Timeman\TimemanServiceBuilder;
 use Psr\Log\LoggerInterface;
@@ -420,6 +421,20 @@ class ServiceBuilder extends AbstractServiceBuilder
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new MessageserviceServiceBuilder(
 
+                $this->core,
+                $this->batch,
+                $this->bulkItemsReader,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function getNoteScope(): NoteServiceBuilder
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new NoteServiceBuilder(
                 $this->core,
                 $this->batch,
                 $this->bulkItemsReader,